Czoo.jl

This project aims to create examples of interfacing Julia code with C-Code

Czoo.free_ptr — Function

free_ptr(ptr; free=true)

Free the given ptr if not C_NULL.

Arguments

ptr the C pointer to be freed

free a conditional flag to prevent an if at the call site

Czoo.add — Function

add(a, b)

Calls C to adds two Ints and returns the result

Czoo.concat — Function

concat(a, b)

Calls C to concat two Strings and returns the result (and frees the memory)

Czoo.cons — Function

cons(a, b)

Calls C to create a list of two Strings and returns the result as a Vector{String} (and frees the memory)

Czoo.CPstring — Type

CPstring

A "Pascal" style string with a length and the uchars

Properties

length::Cint length of the uchars

uchars::Ptr{Cuchar} pointer to the uchars

Czoo.Pstring — Type

Pstring

The Julia version of the CPstring

Properties

length::Int length of the string

uchars::String the string

Constructor

Pstring(cp::CPstring; free=true) - create a Pstring from a CPstring and optionally free the memory

Czoo.pstring — Function

pstring(a)

Calls C to create a CPstring struct and returns a PString (and frees the memory used for the string)

Czoo.ptr_pstring — Function

ptr_pstring(a)

Calls C to create a Ptr{CPstring} struct and returns a PString (and frees the memory of the string and the malloc'd struct)

Czoo.CLinkedPstring — Type

CLinkedPstring

A "Pascal" style string which also links to the "next" in the list

Properties

next::Ptr{CLinkedPstring}

length::Cint

uchars::Ptr{Cuchar}

Czoo.LinkedPstring — Type

LinkedPstring

The Julia version of the CLinkedPstring

Constructor

LinkedPstring(clp::CLinkedPstring; free=true) - create a LinkedPstring (following all the links) from a CLinkedPstring and optionally free the memory

LinkedPstring(pclp::Ptr{CLinkedPstring}; free=true) - create a LinkedPstring (following all the links) from a Ptr{CLinkedPstring} and optionally free the memory

Czoo.linked_pstrings — Function

linked_pstrings(a, b, c)

Calls C to create a Ptr{CLinkedPstring} struct and returns a LinkedPString (and frees the memory of the string and the malloc'd structs)